build: update build

+++ build/use-cases/products/delete.js
@@ -12,14 +12,19 @@ Object.defineProperty(exports, "__esModule", { value: true });
 exports.DeleteProductsUseCase = void 0;
 const errors_1 = require("../../errors");
 class DeleteProductsUseCase {
-    constructor(productsRepo) {
+    constructor(productsRepo, salesRepo) {
         this.productsRepo = productsRepo;
+        this.salesRepo = salesRepo;
     }
     execute(productId) {
         return __awaiter(this, void 0, void 0, function* () {
             const product = yield this.productsRepo.findById(productId);
             if (!product) {
                 throw new errors_1.NotFoundError('Product not found');
+            }
+            const sales = yield this.salesRepo.findByProductId(productId);
+            if (sales) {
+                throw new errors_1.ConflictError('Product has associated sales and cannot be deleted');
             }
             yield Promise.all([
                 this.productsRepo.deleteImages(product.id),
